Rawr.Base: Fix for Issue 19667: * Optimizer showing items that do not require modification - There actually were changes being made. The reforging Id was being brought down in iterations of 56. Placed a check in where reforge id's are loaded in to prevent this (will load them into the proper values first, so they wont be reassigned again later).
git-tfs-id: [https://tfs.codeplex.com/tfs/TFS03]$/Rawr;C58697

+++ Rawr.Base/Items/Item.cs
@@ -1203,7 +1203,7 @@ namespace Rawr
             {
                 if (Reforging != null)
                 {
-                    return (int)Reforging.Id;
+                    return Reforging.Id;
                 }
                 return 0; 
             }
